Add files via upload
This is one of the most simple chatbots you can build with very few lines of code. Of course, if you want a more sophisticated chatbot then it all depends on the scale & vastness of the corpus which we give for training. Hope this helps my fellow friends and Data Science aspirants.